# docs/api/companies_openapi.yml
openapi: 3.0.0
info:
  title: ms-companies
  version: 1.0.0
servers:
  - url: https://ms-companies-dev.organization.com
components:
  securitySchemes:
    noauthAuth:
      type: http
      scheme: noauth
...
paths:
  /v4/companies/size:
    get:
      tags:
        - CompanyModel
      summary: NumberOfCompanies
      security:
        - noauthAuth: []
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{token}}
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Content-Length:
              schema:
                type: integer
                example: '18'
          content:
            application/json:
              schema:
                type: object
              example:
                size: 221
        '404':
          description: NOT FOUND
          headers:
            Content-Type:
              schema:
                type: string
                example: text/html; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '232'
          content:
            text/plain:
              schema:
                type: string
              example: >-
                <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">

                <title>404 Not Found</title>

                <h1>Not Found</h1>

                <p>The requested URL was not found on the server. If you entered
                the URL manually please check your spelling and try again.</p>
  /v4/company/:
    post:
      tags:
        - CompanyModel
      summary: Create
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: prueba
                phone: 9
                email: ''
      security:
        - noauthAuth: []
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{token}}
      responses:
        '201':
          description: CREATED
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Content-Length:
              schema:
                type: integer
                example: '218'
          content:
            application/json:
              schema:
                type: object
              example:
                created: '2022-06-22 21:00:44.992127'
                email: ''
                is_active: true
                modified: '2022-06-22 21:00:44.992145'
                name: prueba
                phone: '9'
                uuid: 9b27941f-aad1-489c-b2e9-4a63ec7ad80a
        '400':
          description: BAD REQUEST
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Content-Length:
              schema:
                type: integer
                example: '37'
          content:
            application/json:
              schema:
                type: object
              examples:
                example-0:
                  summary: Create company - 400 BAD REQ - Empty Body request
                  value:
                    error: Company need a name
                example-1:
                  summary: Create - 400 BAD REQ - Body without name info to request
                  value:
                    error: Company need a name
                example-2:
                  summary: >-
                    Create company - 400 BAD REQ - Body without phone info
                    request
                  value:
                    error: Company need a phone
                example-3:
                  summary: >-
                    Create company - 400 BAD REQ - Body without email info
                    request
                  value:
                    error: Company need a email
...
  /v4/companies/:
    get:
      tags:
        - CompanyModel
      summary: ListAll
      security:
        - noauthAuth: []
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{token}}
        - name: page
          in: query
          schema:
            type: integer
          example: '4'
        - name: showing_range
          in: query
          schema:
            type: integer
          example: '30'
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Content-Length:
              schema:
                type: integer
                example: '7899'
          content:
            application/json:
              schema:
                type: object
              example:
                pages: 8
                resource:
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Iqor
                    phone: '0'
                    uuid: 513937d9-1f86-4162-b5f8-c59a8c9f8c91
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: AMDA
                    phone: '0'
                    uuid: d63ecf24-083b-4d64-83ff-dc999a57d41e
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Rone
                    phone: '0'
                    uuid: 3cc32a00-9bc1-48cb-929c-2cd657c1fd83
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: ExitusCredit
                    phone: '0'
                    uuid: aeaaf15b-7132-4daf-be02-0ed1cfe59048
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: NATUREGANIX
                    phone: '0'
                    uuid: f751d765-cdeb-47d3-95c3-177847ba0036
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: FortemCapital
                    phone: '0'
                    uuid: c185a12e-14fd-4a0b-bf80-4a92f18f1674
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: TexasRibs
                    phone: '0'
                    uuid: e88b4cad-7fa5-4236-9488-91bc0f80086c
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Selmec
                    phone: '0'
                    uuid: 1b60821b-caa4-43c7-b8ee-f4f953f2b3f2
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Pearson
                    phone: '0'
                    uuid: 0facba1d-4208-4add-b65b-8f36155a115d
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Cosco Shipping Lines Mexico
                    phone: '0'
                    uuid: 9345e9ec-813a-4388-ba39-096e258ea709
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Chg-Meridian
                    phone: '0'
                    uuid: 040f8878-6f71-4408-ab49-39ea6dce8b5f
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Tecnológico de Monterrey CSF
                    phone: '0'
                    uuid: f244223d-a371-4be1-95ab-23867f1a1197
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Amazon
                    phone: '0'
                    uuid: 031af7f1-5730-4456-807a-1f10842a1358
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: SubZ3ro
                    phone: '0'
                    uuid: d2fb9ba3-798c-47c3-8a78-27a4baed3d4f
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Comercializadora Bargo S.A de C.V
                    phone: '0'
                    uuid: 2298c20f-b971-45d6-846a-7ecdafa4b1e3
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Tiendas Neto
                    phone: '0'
                    uuid: 7e1a19a4-9979-4bb9-9646-8fc30e6b6b6b
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Grupo Cauda
                    phone: '0'
                    uuid: 27269ac3-1b3a-498b-858f-3c1f0859a390
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Toyota
                    phone: '0'
                    uuid: d3954287-eedb-405f-b663-12574a990a25
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Ottomotores
                    phone: '0'
                    uuid: c288b481-6d56-41e2-93f5-42b0d8aa4ca9
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Cielito Querido Cafe
                    phone: '0'
                    uuid: 5958fdc0-f591-4b68-8200-165b0d551296
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Alto Potencial
                    phone: '0'
                    uuid: fe0550b4-985a-440d-9369-be605265c15d
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: 3M
                    phone: '0'
                    uuid: 793b0eea-cd35-4197-bc75-5da9e7a33324
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Bonafont
                    phone: '0'
                    uuid: c909bb8d-d927-47c3-b630-12602026799c
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: P&G
                    phone: '0'
                    uuid: 65b1cdb4-1d72-44d8-a85e-a7637a872998
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Avery Dennison
                    phone: '0'
                    uuid: f72251b1-6755-40b1-b1a1-d9fa32c0a0ba
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Asofarma
                    phone: '0'
                    uuid: 7377cafd-46b1-46ec-9eec-3d740f00c2b1
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Médica Sur
                    phone: '0'
                    uuid: fa288177-96f4-419b-b20d-c475f90a6ee3
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: YEMA
                    phone: '0'
                    uuid: e1b1e4e7-b5c6-406b-9bbd-cacbbdd40ab7
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Canasta Rosa
                    phone: '0'
                    uuid: ec59e1d4-2482-4d21-8ecd-f41b7f451255
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Procesa MX (Atún Marina Azul)
                    phone: '0'
                    uuid: bfa1054f-65b8-4dc1-834e-3112c6e0d621
        '400':
          description: BAD REQUEST
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Content-Length:
              schema:
                type: integer
                example: '41'
          content:
            application/json:
              schema:
                type: object
              examples:
                example-0:
                  summary: ListAll - 400 BAD REQ - Out of range
                  value:
                    error: list index out of range
                example-1:
                  summary: ListAll - 400 BAD REQ - No Params
                  value:
                    error: >-
                      int() argument must be a string, a bytes-like object or a
                      real number, not 'NoneType'
...
  /v4/companies/sort_by_name/asc:
    get:
      tags:
        - CompanyModel
      summary: SortByNameASC
      security:
        - noauthAuth: []
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{token}}
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: showing_range
          in: query
          schema:
            type: integer
          example: '30'
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Content-Length:
              schema:
                type: integer
                example: '7884'
          content:
            application/json:
              schema:
                type: object
              example:
                pages: 8
                resource:
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: 3M
                    phone: '0'
                    uuid: 793b0eea-cd35-4197-bc75-5da9e7a33324
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: 99 Minutos
                    phone: '0'
                    uuid: b50d239b-44c6-48b9-be34-f3dd52cb9422
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2022-06-22 23:27:08.872186'
                    name: ABSA
                    phone: '0'
                    uuid: cf959795-c7fd-45a8-9c1f-de23a8f2888d
                  - created: '2022-02-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2022-02-02 21:25:37.912000'
                    name: AGNP MTY
                    phone: '0'
                    uuid: 414de945-4783-410b-b77d-47aa36776726
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: AMDA
                    phone: '0'
                    uuid: d63ecf24-083b-4d64-83ff-dc999a57d41e
                  - created: '2022-05-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2022-05-02 21:25:37.912000'
                    name: Aliat university
                    phone: '0'
                    uuid: e652f4df-7a32-452f-a13b-156a0a8afe3f
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Alloy
                    phone: '0'
                    uuid: ce7c341a-035d-4956-9b19-ddc0de7de49a
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Alto Potencial
                    phone: '0'
                    uuid: fe0550b4-985a-440d-9369-be605265c15d
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Alyrgo
                    phone: '0'
                    uuid: 18408172-fef2-41da-9694-722c5568a715
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Amazon
                    phone: '0'
                    uuid: 031af7f1-5730-4456-807a-1f10842a1358
                  - created: '2021-01-14 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-01-14 21:25:37.912000'
                    name: Amazon Apodaca - Zuazua/Agua fría
                    phone: '0'
                    uuid: 076bea7c-db95-4fa2-b22c-72b7c6f41128
                  - created: '2020-12-09 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2020-12-09 21:25:37.912000'
                    name: Amazon Contingencia (No funciona)
                    phone: '0'
                    uuid: 5c837a47-d8a8-437b-a4f3-adf430285759
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DCL1
                    phone: '0'
                    uuid: 22fa9c04-7fe3-4f80-b32f-5028c8dc7531
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DGD1
                    phone: '0'
                    uuid: aa11f82f-88aa-4fd7-b330-5c795222a210
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DGD2
                    phone: '0'
                    uuid: 7db405c3-9d1c-4fb2-9258-78a12037439c
                  - created: '2021-06-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-06-02 21:25:37.912000'
                    name: Amazon DHM1
                    phone: '0'
                    uuid: 5c05c939-ad50-4d09-8c55-31c1e6e10ad1
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DMT1
                    phone: '0'
                    uuid: 02120c4f-f884-49cc-8a01-2e92611e537d
                  - created: '2021-06-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-06-02 21:25:37.912000'
                    name: Amazon DMT3
                    phone: '0'
                    uuid: 020eb478-5f25-4988-875e-5144c3bbab18
                  - created: '2021-07-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-07-02 21:25:37.912000'
                    name: Amazon DMX13
                    phone: '0'
                    uuid: 08a3e62e-5d39-4b6d-adb7-28549561df92
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DMX2
                    phone: '0'
                    uuid: d8333ffb-e9c7-4876-9a0c-6d476047978a
                  - created: '2021-07-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-07-02 21:25:37.912000'
                    name: Amazon DMX25
                    phone: '0'
                    uuid: f251afae-3050-4e7d-b500-343fad1afa90
                  - created: '2021-07-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-07-02 21:25:37.912000'
                    name: Amazon DMX4
                    phone: '0'
                    uuid: ee2b12b8-d40f-4cae-9b89-a23adfeb3168
                  - created: '2021-10-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-10-02 21:25:37.912000'
                    name: Amazon DMX6
                    phone: '0'
                    uuid: 900f3d5c-19af-4313-9a8f-3b30df42b8ae
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DPB1
                    phone: '0'
                    uuid: 5a71c559-c0e2-4689-a2ea-dba838def9fc
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DPB15
                    phone: '0'
                    uuid: 69f3d894-f5f3-434e-ae6a-c0e7c33a1d9d
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DQR1
                    phone: '0'
                    uuid: 28c71abe-c3d2-46b6-b332-c4d26b6b0ee6
                  - created: '2021-08-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-08-02 21:25:37.912000'
                    name: Amazon DSL1
                    phone: '0'
                    uuid: 0a320c93-98e8-4c7f-80a4-8cd1c2189d69
                  - created: '2021-08-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-08-02 21:25:37.912000'
                    name: Amazon DTJ1
                    phone: '0'
                    uuid: 5782988a-273b-4233-8eb9-72db6a571d62
                  - created: '2021-03-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-03-02 21:25:37.912000'
                    name: Amazon DTL1
                    phone: '0'
                    uuid: 0991fb04-a43a-40a1-afea-ea507c5fba15
                  - created: '2021-10-02 21:25:37.912000'
                    email: ''
                    is_active: true
                    modified: '2021-10-02 21:25:37.912000'
                    name: Amazon DUU1
                    phone: '0'
                    uuid: 58ee4919-b30a-49b2-9712-24eeb5046159
...